Split off PeriodLabels sub-component.
